feat(sharing): Refactor sharing functionality with new routers and merge logic
- Introduce new sharing routers for handling share requests, acceptance, linking, and metadata.
- Implement merging of routers to streamline sharing operations.
- Add new methods for managing shared items, including linking shared players and games.
- Enhance share request handling with improved validation and error management.
- Remove the deprecated sharing router file to clean up the codebase.
