+++ packages/api/src/routers/sharing/index.ts
@@ -1,0 +1,14 @@
+import { mergeRouters } from "../../trpc";
+import { shareAcceptanceRouter } from "./share-accept";
+import { shareLinkRouter } from "./share-link";
+import { shareLinkingRouter } from "./share-linking";
+import { shareMetaRouter } from "./share-meta";
+import { shareRequestRouter } from "./share-request";
+
+export const sharingRouter = mergeRouters(
+  shareAcceptanceRouter,
+  shareLinkRouter,
+  shareLinkingRouter,
+  shareMetaRouter,
+  shareRequestRouter,
+);
