+++ packages/api/src/trpc.ts
@@ -83,6 +83,7 @@ export const createCallerFactory = t.createCallerFactory;
  * @see https://trpc.io/docs/router
  */
 export const createTRPCRouter = t.router;
+export const mergeRouters = t.mergeRouters;
 
 /**
  * Middleware for timing procedure execution and adding an artificial delay in development.
